JAMES-2344 QuotaRoot is part of mailbox-api and should not require an implementation project to create instances
