+++ mailbox/store/src/test/java/org/apache/james/mailbox/store/quota/DefaultQuotaRootResolverTest.java
@@ -41,7 +41,7 @@ public class DefaultQuotaRootResolverTest {
     public static final MailboxPath PATH_LIKE = MailboxPath.forUser("benwa", "%");
     public static final MailboxPath MAILBOX_PATH_2 = MailboxPath.forUser("benwa", "test");
     public static final SimpleMailbox MAILBOX_2 = new SimpleMailbox(MAILBOX_PATH_2, 10);
-    public static final QuotaRoot QUOTA_ROOT = QuotaRootImpl.quotaRoot("#private&benwa");
+    public static final QuotaRoot QUOTA_ROOT = QuotaRoot.quotaRoot("#private&benwa");
 
     private DefaultQuotaRootResolver testee;
     private MailboxSessionMapperFactory mockedFactory;
@@ -77,7 +77,7 @@ public class DefaultQuotaRootResolverTest {
 
     @Test(expected = MailboxException.class)
     public void retrieveAssociatedMailboxesShouldThrowWhenQuotaRootNotContainsSeparator2Times() throws Exception {
-        testee.retrieveAssociatedMailboxes(QuotaRootImpl.quotaRoot("#private&be&nwa"), null);
+        testee.retrieveAssociatedMailboxes(QuotaRoot.quotaRoot("#private&be&nwa"), null);
     }
 
 }
